@@ -5,43 +5,39 @@
   "main": "index.js",
   "homepage": "https://eshanchatwal.co.uk",
   "scripts": {
-    "start": "webpack-dev-server --mode-development",
-    "build": "webpack -p"
-    },
+    "start": "webpack-dev-server --mode development",
+    "build": "webpack --mode production"
+  },
   "author": "eshan1799",
   "license": "ISC",
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.11.5",
-    "@babel/runtime": "^7.11.2",
-    "babel-jest": "^26.2.2",
-    "babel-loader": "^8.1.0",
-    "css-loader": "^4.3.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-transform-runtime": "^7.21.0",
+    "@babel/runtime": "^7.21.0",
+    "babel-loader": "^9.1.2",
+    "css-loader": "^6.7.3",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^4.3.0",
-    "jest": "^26.2.2",
-    "sinon": "^9.0.2",
-    "style-loader": "^1.3.0",
-    "svg-url-loader": "^6.0.0",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0"
+    "html-loader": "^4.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "style-loader": "^3.3.2",
+    "svg-url-loader": "^8.0.0",
+    "webpack": "^5.76.3",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.13.2"
   },
   "dependencies": {
-    "@babel/core": "^7.11.1",
-    "@babel/preset-env": "^7.11.0",
-    "@babel/preset-react": "^7.10.4",
-    "bootstrap": "^4.5.3",
-    "favicons-webpack-plugin": "^4.2.0",
-    "react": "^16.13.1",
+    "@babel/core": "^7.21.3",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "bootstrap": "^5.2.3",
+    "favicons-webpack-plugin": "^6.0.0",
+    "react": "^18.2.0",
     "react-anchor-link-smooth-scroll": "^1.0.12",
-    "react-bootstrap": "^1.4.0",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
+    "react-bootstrap": "^2.7.2",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.9.0",
     "react-scrollspy": "^3.4.3",
-    "react-slick": "^0.27.13",
+    "react-slick": "^0.29.0",
     "scroll-background": "^1.0.5",
     "slick-carousel": "^1.8.1"
   }
