@@ -1,20 +1,17 @@
 import React from "react";
-import { Switch, Route } from "react-router-dom";
+import { Router, Routes, Route } from "react-router-dom";
 import { HomePage, Error404 } from './Containers/index/index';
 import "./styles/App.css";
 
-class App extends React.Component {
-  render() {
-    return (
-      <>
-        <div className="App">
-          <Switch>
-            <Route exact path='/' component={ HomePage } />
-            <Route component={ Error404 } />
-          </Switch>
-      </div>
-      </>
-    );
-  }
+const App = () => {
+  return (
+    <div className="App">
+        <Routes>
+          <Route exact path="/" element={ <HomePage/> }/>
+          <Route element={ Error404 }/>
+        </Routes>
+    </div>
+  )
 }
+
 export default App;
